JWT 관련 로직 추가

+++ src/main/java/com/backend/wear/controller/UserController.java
@@ -1,13 +1,16 @@
 package com.backend.wear.controller;
 
+import com.backend.wear.config.jwt.Token;
 import com.backend.wear.dto.blockeduser.BlockedUserResponseDto;
 import com.backend.wear.dto.donation.DonationApplyResponseDto;
 import com.backend.wear.dto.product.ProductRequestDto;
 import com.backend.wear.dto.product.ProductResponseDto;
 import com.backend.wear.dto.user.UserRequestDto;
 import com.backend.wear.dto.user.UserResponseDto;
+import com.backend.wear.service.TokenService;
 import com.backend.wear.service.UserService;
 import jakarta.validation.Valid;
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,41 +20,47 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
+@Slf4j
 @RestController
 @RequestMapping("/api/users")
 public class UserController {
-
-    private static final Logger log = LoggerFactory.getLogger(UserController.class);
-
     private final UserService userService;
+    private final TokenService tokenService;
 
     @Autowired
-    public UserController(UserService userService){
+    public UserController(UserService userService, TokenService tokenService){
         this.userService=userService;
+        this.tokenService=tokenService;
     }
 
     // 마이페이지 사용자 정보
     // /api/users/{userId}
     @GetMapping("/{userId}")
-    public ResponseEntity<?> getMyPageUser(@PathVariable(name="userId") Long userId ,Authentication authentication) throws Exception {
-        UserResponseDto.MyPageDto myPageDto;
+    public ResponseEntity<?> getMyPageUser(@PathVariable(name="userId") Long userId ,@RequestHeader("Authorization") String authorizationHeader)
+            throws Exception {
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
 
         try{
-            myPageDto=userService.getMyPageUserService(userId);
+            UserResponseDto.MyPageDto myPageDto=userService.getMyPageUserService(userId);
+            return ResponseEntity.ok(myPageDto);
         }
         catch (IllegalArgumentException e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
                     .body(e.getMessage());
         }
-        return ResponseEntity.ok(myPageDto);
     }
 
     // 사용자 프로필
     // /api/users/profile/{userId}
     @GetMapping("/profile/{userId}")
-    public ResponseEntity<?> getUserProfile(@PathVariable(name="userId") Long userId) throws Exception{
-        UserResponseDto.ProfileDto profileDto;
+    public ResponseEntity<?> getUserProfile(@PathVariable(name="userId") Long userId, @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
 
+        UserResponseDto.ProfileDto profileDto;
         try{
             profileDto =userService.getUserProfileService(userId);
 
@@ -67,9 +76,13 @@ public class UserController {
     // /api/users/profile/{userId}
     @PutMapping("/profile/{userId}")
     public ResponseEntity<?> updateUserProfile(@PathVariable(name="userId") Long userId,
-                                               @RequestBody @Valid UserRequestDto.ProfileDto profileDto)
+                                               @RequestBody UserRequestDto.ProfileDto profileDto, @RequestHeader("Authorization") String authorizationHeader)
     throws Exception
     {
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             userService.updateUserProfile(userId, profileDto);
             return ResponseEntity.ok().body("회원 정보가 변경되었습니다.");
@@ -82,9 +95,12 @@ public class UserController {
     // 계정 정보
     // /api/users/userInfo/{userId}
     @GetMapping("/userInfo/{userId}")
-    public ResponseEntity<?> getUserInfo(@PathVariable(name="userId") Long userId){
-        UserResponseDto.InfoDto userResponseDto;
+    public ResponseEntity<?> getUserInfo(@PathVariable(name="userId") Long userId, @RequestHeader("Authorization") String authorizationHeader){
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
 
+        UserResponseDto.InfoDto userResponseDto;
         try{
             userResponseDto=userService.getUserInfoService(userId);
         } catch (IllegalArgumentException e) {
@@ -98,7 +114,12 @@ public class UserController {
     // 계정 정보 저장
     // /api/users/userInfo/update/{userId}
     @PutMapping("/userInfo/update/{userId}")
-    public ResponseEntity<?> updateUserInfo(@PathVariable(name="userId") Long userId,@RequestBody @Valid UserRequestDto.InfoDto infodto){
+    public ResponseEntity<?> updateUserInfo(@PathVariable(name="userId") Long userId,@RequestBody UserRequestDto.InfoDto infodto
+    ,@RequestHeader("Authorization") String authorizationHeader){
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             userService.updateUserInfoService(userId, infodto);
             return ResponseEntity.ok().body("사용자 이름이 변경되었습니다.");
@@ -111,8 +132,12 @@ public class UserController {
     // 비밀번호 변경하기
     // /api/users/password/{userId}
     @PutMapping ("/password/{userId}")
-    public ResponseEntity<?> putPassword(@PathVariable(name="userId") Long userId, @RequestBody @Valid UserRequestDto.PasswordDto passwordDto)
-    throws Exception{
+    public ResponseEntity<?> putPassword(@PathVariable(name="userId") Long userId, @RequestBody UserRequestDto.PasswordDto passwordDto,
+                                         @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             userService.updatePassword(userId, passwordDto);
             return ResponseEntity.ok().body("비밀번호가 변경되었습니다.");
@@ -126,7 +151,12 @@ public class UserController {
     // /api/users/wishList/{userId}?pageNumber={}
     @GetMapping("/wishList/{userId}")
     public ResponseEntity<?> getWishList(@PathVariable(name="userId") Long userId,
-                                         @RequestParam(name="pageNumber") Integer pageNumber) throws Exception{
+                                         @RequestParam(name="pageNumber") Integer pageNumber ,
+                                         @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
         try {
             Page <ProductResponseDto.ScreenDto> myWishPage = userService.getMyWishPage(userId, pageNumber);
             return ResponseEntity.ok(myWishPage);
@@ -140,7 +170,11 @@ public class UserController {
     // /api/users/myProducts/onSale/{userId}?pageNumber={}
     @GetMapping("/myProducts/onSale/{userId}")
     public ResponseEntity<?> getMyProductsOnSale(@PathVariable(name="userId") Long userId,
-                                                 @RequestParam(name="pageNumber")Integer pageNumber) throws Exception{
+                                                 @RequestParam(name="pageNumber")Integer pageNumber, @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try{
             Page<ProductResponseDto.MyPageScreenDto> myProductsPage =
                     userService.getMyProductsPage(userId,"onSale", pageNumber);
@@ -155,7 +189,11 @@ public class UserController {
     // /api/users/myProducts/onSale/{userId}
     @PutMapping("/myProducts/onSale/{userId}")
     public ResponseEntity<?> updatePostStatusSoldOut(@PathVariable(name="userId") Long userId,
-                                                 @RequestBody @Valid ProductRequestDto requestDto) throws Exception{
+                                                 @RequestBody @Valid ProductRequestDto requestDto, @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try{
             userService.changePostStatus(userId, requestDto);
             return ResponseEntity.ok().body("상품 판매가 완료되었습니다.");
@@ -169,7 +207,11 @@ public class UserController {
     // /api/users/myProducts/soldOut/{userId}?pageNumber={pageNumber}
     @GetMapping("/myProducts/soldOut/{userId}")
     public ResponseEntity<?> getMyProductsSoldOut(@PathVariable(name="userId") Long userId,
-                                                  @RequestParam(name="pageNumber")Integer pageNumber) throws Exception{
+                                                  @RequestParam(name="pageNumber")Integer pageNumber, @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try{
             Page<ProductResponseDto.MyPageScreenDto> myProductsPage =
                     userService.getMyProductsPage(userId,"soldOut", pageNumber);
@@ -184,7 +226,11 @@ public class UserController {
     // /api/users/myProducts/private/{userId}?pageNumber={}
     @GetMapping("/myProducts/private/{userId}")
     public ResponseEntity<?> getMyProductsPrivate(@PathVariable(name="userId") Long userId,
-                                                  @RequestParam(name="pageNumber")Integer pageNumber) throws Exception{
+                                                  @RequestParam(name="pageNumber")Integer pageNumber,
+                                                  @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
 
         try{
             Page<ProductResponseDto.PrivateDto> myPrivateProductsPage
@@ -199,8 +245,12 @@ public class UserController {
     // 내 기부 내역 불러오기
     // /api/users/myDonations/{userId}?pageNumber={pageNumber}
     @GetMapping ("/myDonations/{userId}")
-    public ResponseEntity<?> getMyDonationApply(@PathVariable(name="userId") Long userId, @RequestParam(name="pageNumber") Integer pageNumber)
-            throws Exception{
+    public ResponseEntity<?> getMyDonationApply(@PathVariable(name="userId") Long userId, @RequestParam(name="pageNumber") Integer pageNumber
+            , @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             Page <DonationApplyResponseDto> myDonationApplyPage
                     =userService.myDonationApplysPage(userId, pageNumber);
@@ -215,9 +265,13 @@ public class UserController {
     // 내 기부 내역 중 기부 완료 된 상품만 불러오기
     // /api/users/complete/{userId}?pageNumber={pageNumber}
     @GetMapping ("/myDonations/complete/{userId}")
-    public ResponseEntity<?> getMyDonationApplyComplete(@PathVariable(name="userId") Long userId, @RequestParam(name="pageNumber") Integer pageNumber)
-            throws Exception{
-        try {;
+    public ResponseEntity<?> getMyDonationApplyComplete(@PathVariable(name="userId") Long userId, @RequestParam(name="pageNumber") Integer pageNumber,
+                                                        @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
+        try {
             Page <DonationApplyResponseDto>  myDonationApplyCompletePage
                     =userService.myDonationApplysCompletePage(userId, pageNumber);
             return ResponseEntity.ok().body(myDonationApplyCompletePage);
@@ -231,8 +285,12 @@ public class UserController {
     // 차단한 사용자 리스트 불러오기
     // /api/users/blockedUsers/{userId}?pageNumber={pageNumber}
     @GetMapping("/blockedUsers/{userId}")
-    public ResponseEntity<?> getBlockedUsersList(@PathVariable(name="userId") Long userId, @RequestParam(name="pageNumber") Integer pageNumber)
-            throws Exception{
+    public ResponseEntity<?> getBlockedUsersList(@PathVariable(name="userId") Long userId, @RequestParam(name="pageNumber") Integer pageNumber,
+                                                 @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             Page<BlockedUserResponseDto> blockedUserPage = userService.getBlockedUsersPage(userId, pageNumber);
             return ResponseEntity.ok().body(blockedUserPage);
@@ -247,7 +305,12 @@ public class UserController {
     // /api/users/blockedUsers/unBlock/{userId}/{blockedUserId}
     @DeleteMapping("/blockedUsers/unBlock/{userId}/{blockedUserId}")
     public ResponseEntity<?> deleteBlockedUser(@PathVariable(name="userId") Long userId,
-                                               @PathVariable(name="blockedUserId")Long blockedUserId) throws Exception{
+                                               @PathVariable(name="blockedUserId")Long blockedUserId,
+                                               @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             userService.deleteBlockedUser(userId,blockedUserId);
             return ResponseEntity.ok().body("차단이 해제되었습니다.");
